refactor: useScrollFade 훅 생성 및 적용
